refactor: use helpers, require instead of assert when possible
